+++ Tests/SentryTests/Integrations/Screenshot/TestSentryScreenShot.swift
@@ -1,0 +1,13 @@
+import Foundation
+
+#if os(iOS) || os(tvOS) || targetEnvironment(macCatalyst)
+class TestSentryScreenshot: SentryScreenshot {
+    
+    var result: [Data]?
+        
+    override func appScreenshots() -> [Data]? {
+        return result
+    }
+    
+}
+#endif