+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fix: User feedback crash (#1766)
+- feat: Attach screenshots for errors (#1751)
 - fix: Remove authenticated pointer stripping for iOS backtraces (#1757)
 - fix: NSURL warning during SDK initialization (#1764)